refactored assertions into separate assertions module
